backend/requirements_new.txt
# 새로운 의존성 추가
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.7
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-dotenv==1.0.0
pyjwt==2.9.0
bcrypt==4.1.3
email-validator==2.2.0
